+++ types/index.d.ts
@@ -80,6 +80,7 @@ export declare const Grid: DefineComponent<{
 	header?: boolean;
 	footer?: boolean;
 	reorder?: boolean;
+	draggableRows?: boolean | ((row: IRow) => boolean);
 	autoRowHeight?: boolean;
 	responsive?: {
 		[key: string]: {